# Drastically optimize pause menu elements